testtest

+++ cyx/common/rabitmq_message.py
@@ -123,6 +123,8 @@ class RabitmqMsg:
             self.__channel__.queue_declare(queue=self.get_real_msg(msg_type), auto_delete=False)
             self.__is_declare__ = True
         self.__channel__.basic_consume(queue=self.get_real_msg(msg_type), on_message_callback=callback, auto_ack=False)
+
+
         try:
             self.__channel__.start_consuming()
         except pika.exceptions.ConnectionClosedByBroker as e:
@@ -218,7 +220,9 @@ class RabitmqMsg:
         raise NotImplemented
 
     def delete(self, item: MessageInfo):
-        self.delete_msg(item)
+        if item.deleted is None or item.deleted==False:
+            self.delete_msg(item)
+            item.deleted=True
 
     def delete_msg(self, item: MessageInfo):
         """
